requirements.txt
# requirements.txt
mysql-connector-python
Pillow